refactor: remove v2 module
