+++ benchmark/src/zipf.rs
@@ -19,8 +19,8 @@
 )]
 
 use crate::TestRunner;
+use firewood::api::{Db as _, Proposal as _};
 use firewood::db::{BatchOp, Db};
-use firewood::v2::api::{Db as _, Proposal as _};
 use log::{debug, trace};
 use pretty_duration::pretty_duration;
 use rand::prelude::*;
