+++ ffi/src/handle.rs
@@ -4,13 +4,13 @@
 use std::num::{NonZeroU64, NonZeroUsize};
 
 use firewood::{
-    db::{Db, DbConfig},
-    manager::RevisionManagerConfig,
-    merkle::Merkle,
-    v2::api::{
+    api::{
         self, ArcDynDbView, Db as _, DbView, FrozenChangeProof, HashKey, HashKeyExt, IntoBatchIter,
         KeyType,
     },
+    db::{Db, DbConfig},
+    manager::RevisionManagerConfig,
+    merkle::Merkle,
 };
 
 use crate::{BatchOp, BorrowedBytes, CView, CreateProposalResult, arc_cache::ArcCache};
