+++ firewood/src/proofs/ser.rs
@@ -16,10 +16,10 @@ use super::{
     types::{ProofNode, ProofType},
 };
 use crate::{
+    api::{FrozenChangeProof, FrozenRangeProof},
     db::BatchOp,
     merkle::{Key, Value},
     proofs::magic::{BATCH_DELETE, BATCH_DELETE_RANGE, BATCH_PUT},
-    v2::api::{FrozenChangeProof, FrozenRangeProof},
 };
 
 impl FrozenRangeProof {
