+++ firewood/src/proofs/tests.rs
@@ -7,7 +7,7 @@ use integer_encoding::VarInt;
 use test_case::test_case;
 
 use super::{header::InvalidHeader, magic, reader::ReadError, types::ProofType};
-use crate::v2::api::FrozenRangeProof;
+use crate::api::FrozenRangeProof;
 
 fn create_valid_range_proof() -> (FrozenRangeProof, Vec<u8>) {
     let merkle = crate::merkle::tests::init_merkle((0u8..=10).map(|k| ([k], [k])));
