--- firewood/src/v2/mod.rs
@@ -1,8 +1,0 @@
-// Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE.md for licensing terms.
-
-/// The public API
-pub mod api;
-
-/// A batch operation and associated types
-mod batch_op;
