+++ fwdctl/src/dump.rs
@@ -1,10 +1,10 @@
 // Copyright (C) 2023, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE.md for licensing terms.
 use clap::Args;
+use firewood::api::{self, Db as _};
 use firewood::db::{Db, DbConfig};
 use firewood::iter::MerkleKeyValueIter;
 use firewood::merkle::{Key, Value};
-use firewood::v2::api::{self, Db as _};
 use firewood_storage::FileIoError;
 use std::borrow::Cow;
 use std::error::Error;
