+++ fwdctl/src/replay.rs
@@ -7,8 +7,8 @@ use std::path::PathBuf;
 use std::time::Instant;
 
 use clap::Args;
+use firewood::api::{self, Db as DbApi};
 use firewood::db::{Db, DbConfig};
-use firewood::v2::api::{self, Db as DbApi};
 use firewood_replay::replay_from_file;
 
 use crate::DatabasePath;
